Replace malfunctional website with better link
The original webdomain is marked as expired and "parked" by the
domain administrator.

the link has been captured from https://jjacks.svbtle.com/commercial-open-source-paradox
which has been repeated in a different blog post by the same author
here:

https://www.coss.community/jj/the-100-billion-commercial-oss-index-cossi-511h

on the latter page there is a screenshot of the now defunct website
which appears to be very similar to the link that this commit is
replacing it with.